+++ src/app/modules/plans/form-plan/form-plan.component.ts
@@ -2,6 +2,8 @@ import {Component, OnDestroy, OnInit} from '@angular/core';
 import {PlanModel} from "../show-plans/plan/plan.model";
 import {NgForm} from "@angular/forms";
 import {PlansService} from "../../../shared/services/plans.service";
+import {MatDialog} from '@angular/material';
+import {ConfirmDialogComponent} from '../../../shared/dialogs/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-form-plan',
@@ -18,7 +20,7 @@ export class FormPlanComponent implements OnInit, OnDestroy {
   private deleted = false;
   private error = '';
 
-  constructor(private plansService: PlansService) { }
+  constructor(private plansService: PlansService, private dialog: MatDialog) { }
 
   resetMessages() {
     this.added = false;
@@ -38,31 +40,37 @@ export class FormPlanComponent implements OnInit, OnDestroy {
       return;
     }
     this.plan.additionalWishes = form.value.additionalWishes;
-    if (this.operation === 'add') {
-      this.plansService.addUserPlan(this.plan).subscribe(
-        (data) => {
-          this.added = true;
-        }, (error) => {
-          this.error = this.plansService.handleError(error);
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, { data: { type: this.operation} });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 'true') {
+        if (this.operation === 'add') {
+          this.plansService.addUserPlan(this.plan).subscribe(
+            (data) => {
+              this.added = true;
+            }, (error) => {
+              this.error = this.plansService.handleError(error);
+            }
+          );
+        } else if (this.operation === 'edit') {
+          this.plansService.editUserPlan(this.plan).subscribe(
+            (data) => {
+              this.edited = true;
+            }, (error) => {
+              this.error = this.plansService.handleError(error);
+            }
+          );
+        } else if (this.operation === 'delete') {
+          this.plansService.deleteUserPlan(this.plan).subscribe(
+            (data) => {
+              this.deleted = true;
+            }, (error) => {
+              this.error = this.plansService.handleError(error);
+            }
+          );
         }
-      );
-    } else if (this.operation === 'edit') {
-      this.plansService.editUserPlan(this.plan).subscribe(
-        (data) => {
-          this.edited = true;
-        }, (error) => {
-          this.error = this.plansService.handleError(error);
-        }
-      );
-    } else if (this.operation === 'delete') {
-      this.plansService.deleteUserPlan(this.plan).subscribe(
-        (data) => {
-          this.deleted = true;
-        }, (error) => {
-          this.error = this.plansService.handleError(error);
-        }
-      );
-    }
+      }
+    });
   }
 
   ngOnDestroy(): void {
